back/pyproject.toml
[tool.poetry]
name = "smartlink-backend"
version = "0.1.0"
description = ""
authors = [
    "Mikael CAPELLE <mikael.capelle@thalesaleniaspace.com>",
    "Maxime Rousselet <maxime.rousselet@thalesaleniaspace.com>"
]
license = "Other/Proprietary License"
readme = "README.md"
packages = [{ include = "tas", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pyhumps = "^3.8.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.1"
typing-extensions = "^4.7.1"
fastapi = "^0.103.0"
sqlalchemy = "^2.0.20"
mariadb = "^1.1.7"
uvicorn = "^0.23.2"
pydantic = "^2.3.0"
tas-dcc-orbit-tools = { version = "^0.3.0", source = "artifactory-tdp" }
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
flake8-black = "^0.3.6"
flake8-pyproject = "^1.2.3"
mypy = "^1.5.1"
pytest = "^7.4.0"
pytest-unordered = "^0.5.2"
pyright = "^1.1.327"
isort = "^5.12.0"
httpx = "^0.24.1"
astropy-stubs = {version = "^0.1.1", source = "artifactory-tdp"}

[[tool.poetry.source]]
name = "artifactory-tdp"
url = "https://artifactory.thalesdigital.io/artifactory/api/pypi/private-pypi-tas-dcc/simple"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
# See https://github.com/PyCQA/pycodestyle/issues/373
extend-ignore = ['E203', 'E231']

[tool.isort]
profile = "black"
multi_line_output = 3
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
known_first_party = ['tas']

[tool.mypy]
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "astropy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "poliastro.*"
ignore_missing_imports = true

[tool.pyright]
reportUnknownVariableType = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false

[tool.pytest.ini_options]
addopts = "-ra -s --ignore=tests/backup"
testpaths = ["tests"]
